old/requirements.txt
evadb
pandas
numpy
torch
torchvision